pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
gunicorn = "^20.1.0"
Flask = "^2.0.3"
pymongo = "^4.3.3"
certifi = "^2022.12.7"
mongomock = "^4.1.2"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.14.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
